Control points and comments
Control points are shown and selectable/movable. Cuves are updated live with points movement. Points are stored in lists and deleted if hide is selected again. However, there seems to be an issue with the bbox scaling. Also it turns out to be a problem that the start and endpoint between two bezier curves are not shown and movable. Also, the changes made to a cuve are lost if hide is slected again and something was moved(scene was recreated)